refactor: extract mocks used in both jobinfo and steptwoform test
