+++ src/__mocks__/components/atoms/typography.tsx
@@ -1,0 +1,8 @@
+interface TypographyProps {
+    children: React.ReactNode;
+    className?: string;
+  }
+
+export const mockTypography = ({ children, className }: TypographyProps) => (
+    <div className={className}>{children}</div>
+);
