+++ src/__mocks__/components/ui/combobox.tsx
@@ -1,0 +1,14 @@
+interface ComboboxProps {
+  label: string;
+  onChange: (value: string) => void;
+  value?: string;
+  error?: string;
+}
+
+export const mockCombobox = ({ label, onChange, value, error }: ComboboxProps) => (
+  <div>
+    <label>{label}</label>
+    <input value={value ?? ''} onChange={(e) => onChange(e.target.value)} aria-label={label} />
+    {error && <div className="error-message">{error}</div>}
+  </div>
+);
