+++ src/__mocks__/components/ui/stepper.tsx
@@ -1,0 +1,7 @@
+interface StepperProps {
+  currentStep: number;
+}
+
+export const mockStepper = ({ currentStep }: StepperProps) => (
+  <div>{`Current step: ${currentStep}`}</div>
+);
