+++ src/__tests__/modules/RegisterFormModule/Section/StepTwoForm.test.tsx
@@ -1,129 +1,9 @@
-import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
+import { setupComponentMocks } from '@/__mocks__/components/setUpComponents';
+setupComponentMocks();
 import '@testing-library/jest-dom';
 import { StepTwoForm } from '@/modules/RegisterFormModule/Section/register-2';
 import { RegisterFormData } from '@/lib/register';
-
-interface ComboboxProps {
-  label: string;
-  onChange: (value: string) => void;
-  value?: string;
-  error?: string;
-}
-
-jest.mock('@/components/ui/combobox', () => ({
-  Combobox: ({ label, onChange, value, error }: ComboboxProps) => (
-    <div>
-      <label>{label}</label>
-      <input value={value ?? ''} onChange={(e) => onChange(e.target.value)} aria-label={label} />
-      {error && <div className="error-message">{error}</div>}
-    </div>
-  ),
-}));
-
-interface ComboboxCheckBoxProps {
-  label: string;
-  onChange: (values: string[]) => void;
-  values?: string[];
-  placeholder?: string;
-  error?: string;
-}
-
-jest.mock('@/components/ui/comboboxCheckbox', () => ({
-  ComboboxCheckBox: ({ label, onChange, values, placeholder, error }: ComboboxCheckBoxProps) => {
-    const displayValues = (values || []).map((v) => v.charAt(0).toUpperCase() + v.slice(1));
-
-    return (
-      <div>
-        <label>{label}</label>
-        <input
-          value={displayValues.join(', ')}
-          onChange={(e) => onChange(e.target.value.split(', '))}
-          placeholder={placeholder ?? ''}
-          aria-label={label}
-        />
-        {error && <div className="error-message">{error}</div>}
-      </div>
-    );
-  },
-}));
-
-interface InputProps {
-  name?: string;
-  label?: string;
-  placeholder?: string;
-  type?: string;
-  value?: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  error?: string;
-  className?: string;
-}
-
-interface TextareaProps {
-  textLabel: string;
-  placeholder?: string;
-  value?: string;
-  onChange: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;
-  error?: string;
-}
-
-interface FileInputProps {
-  textLabel: string;
-  accept?: string;
-  state?: string;
-  value?: string | null;
-  onFileSelect: (file: File) => void;
-  error?: string;
-}
-
-jest.mock('@/components', () => ({
-  Typography: ({ children, className }: { children: React.ReactNode; className?: string }) => (
-    <div className={className}>{children}</div>
-  ),
-  Stepper: ({ currentStep }: { currentStep: number }) => (
-    <div>{`Current step: ${currentStep}`}</div>
-  ),
-  Input: ({ name, label, placeholder, type, value, onChange, error, className }: InputProps) => (
-    <div className={className}>
-      {label && <label htmlFor={name}>{label}</label>}
-      <input
-        id={name}
-        name={name}
-        placeholder={placeholder ?? ''}
-        type={type}
-        value={value ?? ''}
-        onChange={onChange}
-        aria-label={label}
-      />
-      {error && <div className="error-message">{error}</div>}
-    </div>
-  ),
-  Textarea: ({ textLabel, placeholder, value, onChange, error }: TextareaProps) => (
-    <div>
-      <label>{textLabel}</label>
-      <textarea
-        placeholder={placeholder ?? ''}
-        value={value ?? ''}
-        onChange={onChange}
-        aria-label={textLabel}
-      />
-      {error && <div className="error-message">{error}</div>}
-    </div>
-  ),
-  FileInput: ({ textLabel, accept, state, value, onFileSelect, error }: FileInputProps) => (
-    <div>
-      <label>{textLabel}</label>
-      <input
-        type="file"
-        accept={accept}
-        onChange={(e) => e.target.files?.[0] && onFileSelect(e.target.files[0])}
-        aria-label={textLabel}
-      />
-      {state === 'filled' && <span>{value}</span>}
-      {error && <div className="error-message">{error}</div>}
-    </div>
-  ),
-}));
 
 describe('StepTwoForm Component', () => {
   const defaultFormData: RegisterFormData = {
